requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.30.1
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic>=2.8.2
pydantic-settings>=2.3.4
google-genai>=0.6.0
nltk>=3.9.1
scikit-learn>=1.6.0
joblib>=1.4.2
PyPDF2>=3.0.1
httpx>=0.27.0
pytest>=8.2.2
